migrate fireToolNotificationHook to hookSystem
