+++ packages/core/src/hooks/hookSystem.ts
@@ -24,6 +24,7 @@ import type {
   BeforeToolSelectionHookOutput,
   McpToolContext,
 } from './types.js';
+import { NotificationType } from './types.js';
 import type { AggregatedHookResult } from './hookAggregator.js';
 import type {
   GenerateContentParameters,
@@ -33,6 +34,7 @@ import type {
   ToolConfig,
   ToolListUnion,
 } from '@google/genai';
+import type { ToolCallConfirmationDetails } from '../tools/tools.js';
 
 /**
  * Main hook system that coordinates all hook-related functionality
@@ -76,6 +78,99 @@ export interface AfterModelHookResult {
   blocked?: boolean;
   /** Reason for blocking or stopping */
   reason?: string;
+}
+
+/**
+ * Serializable representation of tool confirmation details for hooks.
+ * Excludes function properties like onConfirm that can't be serialized.
+ */
+interface SerializableConfirmationDetails {
+  type: 'edit' | 'exec' | 'mcp' | 'info';
+  title: string;
+  // Edit-specific fields
+  fileName?: string;
+  filePath?: string;
+  fileDiff?: string;
+  originalContent?: string | null;
+  newContent?: string;
+  isModifying?: boolean;
+  // Exec-specific fields
+  command?: string;
+  rootCommand?: string;
+  // MCP-specific fields
+  serverName?: string;
+  toolName?: string;
+  toolDisplayName?: string;
+  // Info-specific fields
+  prompt?: string;
+  urls?: string[];
+}
+
+/**
+ * Converts ToolCallConfirmationDetails to a serializable format for hooks.
+ * Excludes function properties (onConfirm, ideConfirmation) that can't be serialized.
+ */
+function toSerializableDetails(
+  details: ToolCallConfirmationDetails,
+): SerializableConfirmationDetails {
+  const base: SerializableConfirmationDetails = {
+    type: details.type,
+    title: details.title,
+  };
+
+  switch (details.type) {
+    case 'edit':
+      return {
+        ...base,
+        fileName: details.fileName,
+        filePath: details.filePath,
+        fileDiff: details.fileDiff,
+        originalContent: details.originalContent,
+        newContent: details.newContent,
+        isModifying: details.isModifying,
+      };
+    case 'exec':
+      return {
+        ...base,
+        command: details.command,
+        rootCommand: details.rootCommand,
+      };
+    case 'mcp':
+      return {
+        ...base,
+        serverName: details.serverName,
+        toolName: details.toolName,
+        toolDisplayName: details.toolDisplayName,
+      };
+    case 'info':
+      return {
+        ...base,
+        prompt: details.prompt,
+        urls: details.urls,
+      };
+    default:
+      return base;
+  }
+}
+
+/**
+ * Gets the message to display in the notification hook for tool confirmation.
+ */
+function getNotificationMessage(
+  confirmationDetails: ToolCallConfirmationDetails,
+): string {
+  switch (confirmationDetails.type) {
+    case 'edit':
+      return `Tool ${confirmationDetails.title} requires editing`;
+    case 'exec':
+      return `Tool ${confirmationDetails.title} requires execution`;
+    case 'mcp':
+      return `Tool ${confirmationDetails.title} requires MCP`;
+    case 'info':
+      return `Tool ${confirmationDetails.title} requires information`;
+    default:
+      return `Tool requires confirmation`;
+  }
 }
 
 export class HookSystem {
@@ -312,7 +407,7 @@ export class HookSystem {
       );
       return result.finalOutput;
     } catch (error) {
-      debugLogger.debug(`BeforeTool hook failed for ${toolName}:`, error);
+      debugLogger.debug(`BeforeToolEvent failed for ${toolName}:`, error);
       return undefined;
     }
   }
@@ -336,8 +431,28 @@ export class HookSystem {
       );
       return result.finalOutput;
     } catch (error) {
-      debugLogger.debug(`AfterTool hook failed for ${toolName}:`, error);
+      debugLogger.debug(`AfterToolEvent failed for ${toolName}:`, error);
       return undefined;
+    }
+  }
+
+  async fireToolNotificationEvent(
+    confirmationDetails: ToolCallConfirmationDetails,
+  ): Promise<void> {
+    try {
+      const message = getNotificationMessage(confirmationDetails);
+      const serializedDetails = toSerializableDetails(confirmationDetails);
+
+      await this.hookEventHandler.fireNotificationEvent(
+        NotificationType.ToolPermission,
+        message,
+        serializedDetails as unknown as Record<string, unknown>,
+      );
+    } catch (error) {
+      debugLogger.debug(
+        `NotificationEvent failed for ${confirmationDetails.title}:`,
+        error,
+      );
     }
   }
 }
